Fix bug calculating indent
The $tokens[$stackPtr]['length'] is less than
strlen($tokens[$stackPtr]['content']) if the token is part of a multi-
line comment (the extra character(s) in $tokens[$stackPtr]['content']
are from the EOL character(s)). For the purposes of counting line
length, I think these should be ignored, and, in any case, when
calculating $indent by comparing two strings, we need to compare apples
with apples.